+++ letstune/__init__.py
@@ -2,7 +2,6 @@ __all__ = [
     "__version__",
     "Params",
     "ModelParams",
-    "Metric",
     "SimpleTrainer",
     "EpochTrainer",
     "tune",
@@ -11,7 +10,6 @@ __all__ = [
 __version__ = "0.3.0-dev"
 
 
-from .metrics import Metric
 from .params import ModelParams, Params
 from .trainer import EpochTrainer, SimpleTrainer